requirements.txt
# Core
langchain-openai==1.1.12
openai==2.30.0
requests==2.33.0
python-dotenv==1.2.2

# Web dashboard
Flask==3.1.3
Flask-SocketIO==5.6.1

# Voz (ElevenLabs)
elevenlabs[pyaudio]==2.40.0

# Dev / Testing
pytest==9.0.2
